requirements.txt
# Django Core
Django==3.1.7
psycopg2-binary==2.8.6

# Admin
django-3-jet==1.0.8

# Network
gunicorn==20.0.4

# Rest API
djangorestframework==3.12.2
drf-yasg==1.20.0

# DB
pymysql==1.0.2
whitenoise==5.2.0
